pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "voxink"
version = "0.1.0"
description = "Extract timed lyrics from songs automatically - vocal separation + speech recognition + LRC generation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
keywords = ["lyrics", "lrc", "music", "vocal-separation", "transcription", "whisper", "demucs"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "openai-whisper>=20231117",
    "demucs>=4.0.0",
    "torch>=2.0.0",
]

[project.scripts]
voxink = "voxink.cli:main"

[project.urls]
Homepage = "https://github.com/maysh/VoxInk"
Issues = "https://github.com/maysh/VoxInk/issues"
